@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
     <PropertyGroup>
-        <TargetFramework>netcoreapp2.1</TargetFramework>
+        <TargetFramework>netcoreapp3.0</TargetFramework>
         <TypeScriptCompileBlocked>true</TypeScriptCompileBlocked>
         <TypeScriptToolsVersion>Latest</TypeScriptToolsVersion>
         <IsPackable>false</IsPackable>
@@ -10,9 +10,7 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="Microsoft.AspNetCore.App" />
-        <PackageReference Include="Microsoft.AspNetCore.Razor.Design" Version="2.1.2" PrivateAssets="All" />
-        <PackageReference Include="Microsoft.AspNetCore.SpaServices.Extensions" Version="2.1.1" />
+        <PackageReference Include="Microsoft.AspNetCore.SpaServices.Extensions" Version="3.0.0" />
     </ItemGroup>
 
     <ItemGroup>
@@ -23,6 +21,7 @@
 
     <ItemGroup>
       <ProjectReference Include="..\AppleMusicApi\AppleMusicApi.csproj" />
+      <ProjectReference Include="..\LastApple.Persistence\LastApple.Persistence.csproj" />
       <ProjectReference Include="..\LastApple\LastApple.csproj" />
       <ProjectReference Include="..\LastfmApi\LastfmApi.csproj" />
     </ItemGroup>
